pyproject.toml
[project]
name = "dify-zhipu-plugin"
version = "0.1.2"
description = "Dify Zhipu Coding Plan plugin with selected GLM and embedding models"
readme = "README.md"
requires-python = ">=3.12"

# uv pip compile pyproject.toml -o ./requirements.txt
dependencies = [
    "dify_plugin>=0.5.0",
    "zai-sdk>=0.0.3.1",
    "pydantic==2.8.2",
    "sniffio==1.3.1",
]

# uv run black . -C -l 100 && uv run ruff check --fix
[dependency-groups]
dev = []
